refactor(cli): extract artifact generation service
